scraperpy/requirements.txt
Scrapy>=2.0,<3.0
psycopg2-binary>=2.8
scrapy-playwright
